增加排序功能 增加Contact us页面 修复设备名跟随空格问题

+++ itemtouchhelperextension/src/main/java/com/inledco/itemtouchhelperextension/Extension.java
@@ -4,7 +4,7 @@ import android.view.View;
 
 public interface Extension
 {
-    float getActionWidth ();
+    int getActionWidth ();
 
     View getContentView();
 }
